wasm: Don't use nested WebWorker for the fetch thread.
In Safari, Workers can't create new workers. Move the fetch requests to the
main mupdf-worker thread instead.

This code is simpler because we don't need to pass messages to a separate
fetching thread, but we also run the risk of delaying a prefetch request if the
worker thread is busy when a block finishes fetching. This should not matter
too much in practice.