No more cookies.
The "cookie" canceling a job support was never used to good effect in the
viewer. It adds a significant amount of build and runtime complexity, requiring
both building with experimental pthread support in emscripten and proper cross
origin headers from the web server.

Since we don't use it, remove it for now.